Time: 2 ms (38.55%), Space: 11.7 MB (54.93%) - LeetHub

+++ 2819-remove-trailing-zeros-from-a-string/2819-remove-trailing-zeros-from-a-string.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     string removeTrailingZeros(string num) {
-        int numLength = num.length();
-        int lastZeroIndex = -1;
+        unsigned short numLength = num.length();
+        unsigned short lastZeroIndex = -1;
 
-        for (int i = 0; i < numLength; ++i)
+        for (unsigned short i = 0; i < numLength; ++i)
             if (num[i] == '0' && (i != 0 && num[i - 1] != '0'))
                 lastZeroIndex = i;
         
@@ -12,7 +12,7 @@ public:
             lastZeroIndex = -1;
 
         if (lastZeroIndex != -1)
-            for (int i = lastZeroIndex; i < numLength; ++i)
+            for (unsigned short i = lastZeroIndex; i < numLength; ++i)
                 num.pop_back();
         
         return num;
